IGNITE-11731: Fix for Linux

+++ modules/platforms/cpp/core/include/ignite/impl/ignite_environment.h
@@ -342,7 +342,7 @@ namespace ignite
              * @param type Operation type.
              * @param val Value.
              */
-            friend long long IGNITE_CALL InLongOutLong(void* target, int type, long long val);
+            friend int64_t IGNITE_CALL InLongOutLong(void* target, int type, int64_t val);
 
         private:
             /** Node configuration. */
